requirements.txt
numpy~=1.26.0
PyYAML==6.0.1